requirements.txt
requests==2.32.5
PyYAML==6.0.3
snakeviz==2.2.2
pytest==9.0.2
pytest-cov==7.0.0
pandas==3.0.1
beautifulsoup4==4.14.3
pyarrow==23.0.1